move emulator-based tests to integration
