+++ integration/utils/emulator_client.go
@@ -1,4 +1,4 @@
-package module
+package utils
 
 import (
 	"context"
